updated kill script

+++ jenkins/scripts/kill.sh
@@ -4,4 +4,5 @@ echo 'The following command terminates the "npm start" process using its PID'
 echo '(written to ".pidfile"), all of which were conducted when either'
 echo '"deliver-for-development.sh" or "deliver-for-development.sh" was executed.'
 set -x
-kill $(cat .pidfile)
+if pgrep $(cat .pidfile); then kill $(cat .pidfile); fi
+# kill $(cat .pidfile)
